Rename /lib/ensureTenantContext.js to tenant.js
