+++ routes/user.js
@@ -1,11 +1,11 @@
 var express = require('express');
 var passport = require('passport');
 var ensureLoggedIn = require('connect-ensure-login').ensureLoggedIn();
-var ensureTenantContext = require('../lib/ensureTenantContext').ensureTenant();
+var ensureTenant = require('../lib/tenant').ensureTenant();
 var router = express.Router();
 
 /* GET user profile. */
-router.get('/', ensureLoggedIn, ensureTenantContext, function(req, res, next) {
+router.get('/', ensureLoggedIn, ensureTenant, function(req, res, next) {
 
   // The host, which may contain tenant (e.g. tenant1.yourcompany.com)
   var host= req.get('host');
